Pre-check YouTube URLs in the archive before downloading
This tries to stop single video downloads that are in the archive
file from triggering any actual HTTP requests. This became an
issue during the playlist download capability being broken by the
disable_polymer option; as a workaround, I put together download
batch files using a regex-manipulated copy of the playlist page
DOM, so each batch file would run a download for each single video
in a playlist rather than having the program fetch the playlist.

The major problem is that even if a video ID is in the archive,
the video info page must still be downloaded before it is ever
checked within the archive for rejection. By checking the video ID
before anything is downloaded, the rejection happens much faster
and no unnecessary HTTP requests are sent out.

This should be extended in the future for other services that do
not require a page download to retrieve the video ID.

This uses a regex match to check for a youtube.com URL and should
basically be a no-op for all other sites.